requirements.txt
amqp==2.2.2
ansible==2.4.3.0
appnope==0.1.0
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
cryptography==2.3
decorator==4.1.2
Django==1.11
django-cas-client==1.3.0
django-extensions==1.9.8
djangorestframework==3.8.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
influxdb==5.0.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
paramiko==2.4.1
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pymongo==3.6.0
PyMySQL==0.9.2
PyNaCl==1.2.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
xlrd==1.1.0
XlsxWriter==1.0.2
